Struct os1::debug::Debug

pub struct Debug;

A struct to write data to the console port

Methods

impl Debug

fn write_bytes(&self, bytes: &[u8])

Wait for the port, then write the given array of bytes