Module os1::fs::ofs::fs

A module for OFS abstractions

Reexports

use core::prelude::v1::*;

use alloc::arc::Arc;

use core::cmp::min;

use core::mem;

use core::ptr::copy;

use sync::Semaphore;

use string::String;

use io::block::{BlockDevice, BlockDataBuffer};

use io::ide::SECTOR_SIZE;

use super::hw::*;

use super::super::error::Error;

Structs

File
A handle on the file for all needed operations.

OFS
The OFS interface

OFSHandle
A safe handle on the file system for all needed operations.