Struct os1::fs::ofs::fs::File

pub struct File<B: BlockDevice> {
    inode_num: usize,
    inode: Inode,
    offset: usize,
    offset_dnode: usize,
    ofs: Arc<Semaphore<OFS<B>>>,
}

A handle on the file for all needed operations.

Fields

inode_num

inode

offset

offset_dnode

ofs

Methods

impl<B: BlockDevice> File<B>

fn get_filename(&self) -> String

Return the filename

fn read_part(&mut self, buf: &mut BlockDataBuffer) -> usize

Read from the file at the file offset into the buffer at the buffer offset. This method will not overflow the buffer or read past the end of the file, but it might not read as much as possible from the file, even if the buffer is not full. This updates both the file offset and the buffer offset.

fn write_part(&mut self, bytes: usize, buf: &mut BlockDataBuffer) -> usize

Write from the buffer at the buffer offset into the file at the file offset, overwriting any exisiting content. This method may extend the length of the file. This method might not write as much as possible from the file. This updates both the file offset and the buffer offset.

fn seek(&mut self, offset: usize)

Seek into the file to the given offset.

fn read(&mut self, buf: &mut BlockDataBuffer) -> usize

Fill the buffer starting at the buffer offset from the file starting at the file offset. This reads as much as possible from the file without overflowing the buffer or reading past the EOF. This updates both the file and buffer offsets.

fn write(&mut self, bytes: usize, buf: &mut BlockDataBuffer)

Write bytes bytes from the buffer at the buffer offset to the file at the file offset, overwriting any existing content at the offset. This will increase the length of the file if necessary. This updates both the file and buffer offsets.