Struct os1::fs::ofs::fs::OFSHandle

pub struct OFSHandle<B: BlockDevice> {
    fs: Arc<Semaphore<OFS<B>>>,
}

A safe handle on the file system for all needed operations.

Fields

fs

Methods

impl<B: BlockDevice> OFSHandle<B>

fn new(device: B) -> OFSHandle<B>

Create a new handle on the fs using the device

NOTE: we do not have to lock the device now, because ownership transfers to the handle :D

fn open(&mut self, inode: usize) -> Result<File<B>, Error>

Open the file with the given inode number and return a handle to it.

fn link(&mut self, a: usize, b: usize) -> Option<Error>

Create a link (directed edge) from file a to file b. a and b are the inode number of the files.

fn unlink(&mut self, a: usize, b: usize) -> Option<Error>

Remove a link (directed edge) from file a to file b. a and b are the inode number of the files.

fn stat(&self, a: usize) -> Option<Inode>

Return metadata for the file with inode a or None if the file does not exist

fn new_file(&mut self) -> Result<usize, Error>

Create a new file and a link from a to it. a is the inode number of the file. Return the inode number of the new file.

fn delete_file(&mut self, a: usize) -> Option<Error>

Delete file a. a is the inode number of the file.