Module os1::interrupts::idt

This module contains IDT stuff

Reexports

use core::prelude::v1::*;

Structs

IDTDescr
An entry in the IDT

TableDescriptor
A descriptor for use in with instructions like lgdt or lidt

Statics

idt

kernelCodeSeg

Functions

add_interrupt_handler
Register an interrupt handler for the given IRQ number

add_trap_handler
Register a trap handler for the given trap number with the given privelege level