Module os1::io::block

A module for accessing hard disks

Reexports

use core::prelude::v1::*;

use alloc::heap;

use core::cmp::min;

use core::ptr::copy;

use core::mem;

Structs

BlockDataBuffer
A data structure for use with block devices The buffer has an internal pointer that can be used for conveniently performing sequential writes or reads to the buffer. Also, implementors of this trait must also implement Drop, preventing memory leaks.

Traits

BlockDevice
An abstraction over block devices