Struct os1::io::block::BlockDataBuffer

pub struct BlockDataBuffer {
    buf: *mut u8,
    size: usize,
    offset: usize,
}

A data structure for use with block devices The buffer has an internal pointer that can be used for conveniently performing sequential writes or reads to the buffer. Also, implementors of this trait must also implement Drop, preventing memory leaks.

Fields

buf

size

offset

Methods

impl BlockDataBuffer

fn new(size: usize) -> BlockDataBuffer

Create a new buffer with the given capacity

fn offset(&self) -> usize

Get the internal offset in bytes into this buffer

fn set_offset(&mut self, offset: usize)

Set the internal offset in bytes into this buffer

fn size(&self) -> usize

Get the size of this buffer in bytes

unsafe fn get_ptr<T>(&self, offset: usize) -> *mut T

Get a pointer of the given type to the offset in the buffer. The offset is relative to the size of T

Panics

Panics if the offset is out of bounds

unsafe fn get_ref<T>(&self, offset: usize) -> &T

Get a reference of the given type to the offset in the buffer. The offset is relative to the size of T

Panics

Panics if the offset is out of bounds

unsafe fn get_ref_mut<T>(&self, offset: usize) -> &mut T

Get a mutable reference of the given type to the offset in the buffer. The offset is relative to the size of T

Panics

Panics if the offset is out of bounds