Trait os1::io::block::BlockDevice

pub trait BlockDevice {
    fn get_block_size(&self) -> usize;
    fn read_block(&mut self, block_num: usize, buffer: &mut BlockDataBuffer);
    fn write_block(&mut self, block_num: usize, buffer: &BlockDataBuffer);

    fn read(&mut self, offset: usize, buffer: &mut BlockDataBuffer) -> usize { ... }
    fn write(&mut self, offset: usize, buffer: &mut BlockDataBuffer) -> usize { ... }
    fn read_fully(&mut self, offset: usize, buffer: &mut BlockDataBuffer) { ... }
    fn write_fully(&mut self, offset: usize, buffer: &mut BlockDataBuffer) { ... }
    fn read_exactly(&mut self, offset: usize, bytes: usize, buffer: &mut BlockDataBuffer) { ... }
    fn write_exactly(&mut self, offset: usize, bytes: usize, buffer: &mut BlockDataBuffer) { ... }
}

An abstraction over block devices

Required Methods

fn get_block_size(&self) -> usize

Get the block size of this block device in bytes

fn read_block(&mut self, block_num: usize, buffer: &mut BlockDataBuffer)

Read the given block into the buffer

Panics

The method panics if there is not enough space in the buffer

fn write_block(&mut self, block_num: usize, buffer: &BlockDataBuffer)

Write the given block from the buffer

Provided Methods

fn read(&mut self, offset: usize, buffer: &mut BlockDataBuffer) -> usize

Read from the block device at offset into the given buffer starting at the buffer's internal offset. This method will read no more data than will fit into the remaining space in the buffer, but it may also read less. It will update the buffer's offset, and return the number of bytes read.

fn write(&mut self, offset: usize, buffer: &mut BlockDataBuffer) -> usize

Write from the buffer to the disk starting at the buffer's internal offset. This may not write the whole buffer. This will update the buffer's offset, and return the number of bytes written.

fn read_fully(&mut self, offset: usize, buffer: &mut BlockDataBuffer)

Read from the block device at offset into the given buffer starting at the buffer's internal offset. This method will fill the remaining space in the buffer. It will update the buffer's offset.

Panics

The method panics if there is not enough space in the buffer

fn write_fully(&mut self, offset: usize, buffer: &mut BlockDataBuffer)

Write from the buffer to the disk starting at the buffer's internal offset. This will write the whole buffer. This will update the buffer's offset.

fn read_exactly(&mut self, offset: usize, bytes: usize, buffer: &mut BlockDataBuffer)

Read bytes bytes from the block device at offset into the given buffer starting at the buffer's internal offset. This method will fill the remaining space in the buffer. It will update the buffer's offset.

Panics

The method panics if there is not enough space in the buffer

fn write_exactly(&mut self, offset: usize, bytes: usize, buffer: &mut BlockDataBuffer)

Write bytes bytes to the block device at offset from the given buffer starting at the buffer's internal offset. This method will overwrite the existing disk contents. It will update the buffer's offset.