Struct os1::memory::heap::Block

struct Block;

A struct representing a heap block.

It is empty because, by nature, heap blocks are variable size. Rather thsn give the compiler a hard time, we can just a block fixed size and write methods that are aware of the situation. Then, we just need to cast an address into a block pointer. This is a very C-ish solution, but it makes life easy because a heap implementation usually has a lot of pointer arithmetic.

Methods

impl Block

unsafe fn get_head(&self) -> usize

Returns the first word of the block, which will contain the size if this is a free block

unsafe fn set_head(&mut self, head: usize)

Sets the header to the given value

unsafe fn get_foot(&self) -> usize

Returns the last word of the block, which will contain the size if this is a free block

unsafe fn set_foot(&mut self, foot: usize)

Sets the footer to the given value

unsafe fn get_free_bits(&self) -> u8

Gets the 4 free bits of the block

unsafe fn set_next(&mut self, next: *mut Block)

Set the forward pointer (but not the free bits)

unsafe fn set_prev(&mut self, prev: *mut Block)

Set the backward pointer

unsafe fn is_free(&self) -> bool

Returns true if the block is a valid free block. This method does a lot of sanity checiking

unsafe fn mark_free(&mut self)

Set the free bits

unsafe fn mark_used(&mut self)

Clear the free bits. The block should already be removed from the free list

unsafe fn get_size(&self) -> usize

Return the size of the block (user-usable data size)

unsafe fn set_size(&mut self, size: usize)

Set the size of block

unsafe fn get_contiguous_next(&self) -> *mut Block

Returns the heap block immediately following this one

unsafe fn get_contiguous_prev(&self) -> *mut Block

Returns the heap block immediately preceding this one or null if there is no valid previous block.

unsafe fn get_free_next(&self) -> *mut Block

Returns the forward ptr of the block (excluding the free bits)

unsafe fn get_free_prev(&self) -> *mut Block

Returns the backward ptr of the block

unsafe fn remove(&mut self)

Remove from free list. Clear forward/backward pointers

unsafe fn insert(&mut self)

Add to head of free list

unsafe fn merge_with_next(&mut self)

Merge this block with the next block. Removes the second block from the free list before merging

unsafe fn split(&mut self, size: usize)

Split the block so that it is the given size. Insert the new block into the free list. The block must be large enough to split (>= 2*BLOCK_ALIGN). The block must be free. The new size must be a multiple of BLOCK_ALIGN.

unsafe fn is_match(&self, size: usize, align: usize) -> (bool, usize)

Returns true if this block matches the size and alignment. Returns the size at which the block should be split to obtain an aligned block; the second return value is meaningless if we return false.