Module os1::memory::physmem

Physical memory management

Reexports

use core::prelude::v1::*;

use alloc::boxed::Box;

use core::ops::{Index, IndexMut};

use super::super::interrupts::{on, off};

use super::super::process::CURRENT_PROCESS;

use super::super::static_linked_list::StaticLinkedList;

use super::regionmap::RegionMap;

Structs

Frame
A physical memory frame

FrameInfo
Frame info ``` When free 31 20 0 [nnnnnnnnnnnnnnnnnnnn 1] -- next free index -- free bit

FrameInfoSection
Should take up 4MiB on a 32-bit machine. There is one frame info for each frame, associated by frame index

SharedFrameInfo
A struct to keep track of information related to shared physical frames.

Statics

FRAME_INFO
Array of FrameInfo. This is a pointer to the region of memory used to hold physical memory allocation metadata.

FREE_FRAMES
The index of the first free frame in a LIFO list. 0 => empty list

Functions

init
Initialize physical memory frames using the rest of physical memory. This function detects all available physical memory.