Struct os1::memory::physmem::FrameInfo

pub struct FrameInfo {
    info: usize,
}

Frame info ``` When free 31 20 0 [nnnnnnnnnnnnnnnnnnnn 1] -- next free index -- free bit

When allocated 31 30 0 [ssssssssssssssssssssssssssssss 0] -- ptr to shared frame info -- free bit | +-- PD bit ```

Fields

info

Methods

impl FrameInfo

fn alloc(&mut self)

Allocate the frame referred to by this FrameInfo. NOTE: this should only be called on the first frame in the free list

fn free(&mut self)

Free the frame referred to by this FrameInfo if this is the last sharere; otherwise, just remove this process's SharedFrameInfo.

fn get_index(&self) -> usize

Get the index of the Frame

fn set_free(&mut self, free: bool)

Set the free bit of this FrameInfo to 1 if free is true; else 0

fn get_next_free(&self) -> usize

Get the index of the next free frame

fn set_next_free(&mut self, next: usize)

Set the index of the next free frame

fn has_shared_info(&self) -> bool

Returns true if this frame has shared frame info.

NOTE: this does not mean the frame is shared yet.

fn get_shared_info(&self) -> Option<&mut SharedFrameInfo>

Get the SharedFrameInfo of this frame

fn set_shared_info(&mut self, sfi: *mut SharedFrameInfo)

Set the SharedFrameInfo of this frame

fn clear_shared_info(&mut self)

Remove the SharedFrameInfo ptr of the frame