Module os1::memory::regionmap

Module for detecting available physical memory

Get memory map at boot time and use it construct RegionMaps. The BIOS calls needed have to run before we switch to protected mode,

Reexports

use core::prelude::v1::*;

use core::iter::Iterator;

use super::super::vec::Vec;

use super::physmem::Frame;

Structs

MemoryListEntry
Represents an entry in the list of memory regions generated by the E820 BIOS call.

Region
Represents a region of memory

RegionMap
Represents a map of memory regions

Statics

memory_map

memory_map_count

Functions

find
Find the length and type of the most restrictive mapping for addr based on the E820 BIOS call

find_next
Find the beginning address of the next region of memory based on the E820 BIOS call