Module os1::memory::vm

A module for virtual memory management

The virtual address space of a process consists of all addresses for 0xD0_0000-0xFFFF_FFFF. When a page fault occurs (in kernel or user mode), if the faulting address is in this range, the fault is considered an error, and the process is killed. This is because the first 8MiB are shared, direct-mapped memory and should never produce a page fault. The next 4MiB are mapped to the processes page directory, so if there is a fault in this range, it is fatal, since it means the paging structures have been corrupt. The 13MiB is claimed by the kernel for use by the kmap function, which temporarily maps a page meant to reside in another process's address space. This is needed for creating a new process. Luckly, with Rust, segfaults should be extremely rare.

When a page fault occurs for a legal address, the fault handler checks the following cases:

Is the page paged out to disk? If so, swap it back in

Is the page marked present but read-only? If so, this page is COW, so clone it and mark the new page read/write

Else, allocate a new frame a map the page to it

Reexports

use core::prelude::v1::*;

use super::super::interrupts::add_trap_handler;

use super::super::machine::page_fault_handler;

use super::super::static_linked_list::StaticLinkedList;

use super::physmem::Frame;

use self::structs::{VMTable, PagingEntry};

Modules

addr_space
A module process address spaces

structs
A module containing useful structs to abstract paging structures

Structs

AddressSpace
The address space of a single process

Statics

KMAP_ADDRESS
The beginning of kmap memory

NUM_SHARED
The number of shared PDEs (for convenience)

PD_ADDRESS
The virtual address to which the PD is mapped

SHARED_PDES
A list of shared PDEs direct mapping the beginning of memory

USER_ADDRESS
The beginning of user memory

VMM_ON
Is VM on?

Functions

init
Initialize virtual memory management but do not turn on VM.

init_shared_pdes
Create n PDEs to direct map the memory start from the ith page. unsafe because it deals with raw pointers and static muts.

vmm_page_fault
The Rust-side code of the page fault handler.