Struct os1::memory::vm::addr_space::AddressSpace

pub struct AddressSpace {
    page_dir: usize,
    kmap_index: u8,
    lock: StaticSemaphore,
    req_pid: usize,
    req_paddr: usize,
    req_bar: Barrier,
    req_wait: Event,
}

The address space of a single process

Fields

page_dir

paddr of this addr space's PD

kmap_index

the index of the first un-kmapped page

lock

a lock for the address space

NOTE: to prevent deadlocks, ALWAYS acquire this lock before locks on frames.

req_pid

The PID of the process the current share request is going to

req_paddr

The frame the current share request is for

req_bar

A barrier for page-sharing

req_wait

A waiting list for other processes who are expecting this process to share with them

Methods

impl AddressSpace

fn new() -> AddressSpace

Create a new address space and set up the PD.

fn map(&mut self, phys: usize, virt: usize, lock: bool)

Map virt to phys in the address space. The method first tries to acquire the address space lock if lock is true.

NOTE: should only be called on the current address space because it assumes that the PD is at PD_ADDRESS

fn kmap(&mut self, paddr: usize) -> &mut Frame

Map the given paddr for temporary use by the kernel and return a mut reference to the frame. NOTE: should only be called on the current address space because it assumes that the PD is at PD_ADDRESS

fn unmap(&mut self, virt: usize, lock: bool)

Remove any mapping for this virtual address. The method first tries to acquire the address space lock if lock is true.

NOTE: should only be called on the current address space because it assumes that the PD is at PD_ADDRESS

fn v_to_p(&mut self, virt: usize, lock: bool) -> Option<usize>

Returns the current physical address mapped to the given virtual address or None if it is not mapped. The method first tries to acquire the address space lock if lock is true.

NOTE: this has to run while the addresss space is active.

fn activate(&mut self)

Activate the current address space and turn on VM if needed

fn request_share(&mut self, pid: usize, vaddr: usize) -> bool

Send a page-share request to the process with PID pid for the frame mapped to vaddr in this address space.

Returns true if the request succeeded, and false if there was an error (e.g. the process has already died).

Blocks until the request is accepted.

NOTE: this has to run while the addresss space is active.

fn accept_share(&mut self, pid: usize, vaddr: usize) -> bool

Creates a mapping in this process's address space for the first page-share request from the process with the given PID to the given page.

Returns true if success; false otherwise.

Blocks until the request is completed.

NOTE: This should run inside this processes address space

fn clear(&mut self)

Remove all non-kernel mappings in this address space. NOTE: must run while this address space is active