Struct os1::memory::vm::structs::PagingEntry

pub struct PagingEntry {
    entry: usize,
}

A single entry in a page directory or table 31 12 9 0 [ 000 flags ]

Fields

entry

Methods

impl PagingEntry

const fn new() -> PagingEntry

Get a blank entry

fn set_present(&mut self, value: bool)

Set the present bit. true = present; false = not present

fn set_read_write(&mut self, value: bool)

Set the read/write bit. true = read/write, false = read-only

fn set_privelege_level(&mut self, value: bool)

Set the privelege level bit. true = all, false = kernel mode only

fn set_caching(&mut self, value: bool)

Set caching bit. true = write-through, false = write-back

fn set_flag(&mut self, index: u8, value: bool)

Set the index-th flag to value. true = 1; false = 0

fn set_address(&mut self, address: usize)

Set the upper 20 bits of the entry to the upper 20 bits of address

fn is_flag(&self, index: u8) -> bool

Return true if the index-th bit is set

fn get_address(&self) -> usize

Return the upper 20-bits of the entry

fn free(&mut self, dealloc: bool)

Free the frame pointed to if dealloc and mark this entry not present.