Struct os1::process::proc_table::ProcessTable

pub struct ProcessTable {
    pt_head: *mut ProcessTableNode,
    first_pid: usize,
    capacity: usize,
    size: usize,
}

A process table implemented as an indexable deque

Fields

pt_head

first_pid

capacity

size

Methods

impl ProcessTable

const fn new() -> ProcessTable

fn push(&mut self, p: *mut Process)

Add the new process to the end of the Table. Since processes are created with sequential PIDs, we already know the PID.

fn remove(&mut self, pid: usize)

Void the entry in the table. This should be done just before deallocating the process.

fn get(&self, pid: usize) -> Option<*mut Process>

Get the process with the matching PID

fn free(&mut self)

Free as many ProcessTableNodes as possible

fn get_node(&self, pid: usize) -> &'static ProcessTableNode

Get the node containing the given PID

fn get_node_mut(&mut self, pid: usize) -> &'static mut ProcessTableNode

Get the node containing the given PID

fn get_last_node(&mut self) -> &'static mut ProcessTableNode

Get the last node in the table