Module os1::static_linked_list

A doubly-linked list with owned nodes modified slightly from the std library. This version can be used in statics and has a const constructor.

The StaticLinkedList allows pushing and popping elements at either end and is thus efficiently usable as a double-ended queue.

Reexports

use core::prelude::v1::*;

use alloc::boxed::Box;

use core::cmp::Ordering;

use core::fmt;

use core::hash::{Hasher, Hash};

use core::iter::FromIterator;

use core::mem;

use core::ptr;

Structs

IntoIter
An iterator over mutable references to the items of a StaticLinkedList.

Iter
An iterator over references to the items of a StaticLinkedList.

IterMut
An iterator over mutable references to the items of a StaticLinkedList.

Node

Rawlink

StaticLinkedList
A doubly-linked list.

Functions

link_no_prev
Clear the .prev field on next, then return Some(next)

Type Definitions

Link