Struct os1::sync::event::Event

pub struct Event {
    status: StaticSemaphore,
}

An Event data structure.

Processes calling wait will block until a process calls notify.

Fields

status

Methods

impl Event

const fn new() -> Event

Creates a new Event

fn wait(&mut self)

Block this process until notify is called

fn notify(&mut self)

Wake up all processes that called wait on this event

fn reset(&mut self)

Reset the event before more processes can wait