Struct os1::sync::semaphore::StaticSemaphore

pub struct StaticSemaphore {
    count: UnsafeCell<AtomicIsize>,
    queue: UnsafeCell<ProcessQueue>,
}

StaticSemaphore is a semaphore implementation that can be used in statics. It has a const constructor.

Fields

count

queue

Methods

impl StaticSemaphore

const fn new(i: isize) -> StaticSemaphore

Create a new StaticSemaphore with the given count

fn down(&self)

Acquire

fn up(&self)

Release

fn destroy(&self)

Clean up. Cannot implement Drop here because we want to be able to create a static semaphore.