Struct os1::vga::window::Window

pub struct Window {
    height: usize,
    width: usize,
    pos: (usize, usize),
    cursor: (usize, usize),
    vga: VGA,
}

The abstration of a window.

A window has boundaries. Drawing outside those boundaries will not render any text on the screen.

Furthermore, a window can draw a string, rather than just a single character.

Fields

height

width

pos

cursor

vga

Methods

impl Window

fn new(w: usize, h: usize, (row, col): (usize, usize)) -> Window

Create a new Window.

fn clear_screen()

A static method to clear the screen.

fn paint(&mut self)

Paint the window with the background color

fn put_char(&mut self, ch: char)

Draw the character at the cursor relative to the corner of window if the character is inside the window.

fn put_str(&mut self, s: &str)

Render the string at the cursor using word wrapping.

fn set_fg(&mut self, color: Color)

Set the foreground color

fn set_bg(&mut self, color: Color)

Set the background color

fn set_cursor(&mut self, (crow, ccol): (usize, usize))

Set the cursor position